Move hardware writes to Cmm.
Writes in hardware transaction code paths are moved to Cmm and simply
write to the TVar.  We have no record of the write set, so we must also
make retry behave like an abort.  Eventually we will support recording a
summary of the write set in constant space to support retry.