Remove STM atomic invariants.
Invariants are not easily supported with HTM and have not been used (to
our knowledge).  We remove the feature and all related functions, but
leave the `check#` primop inplace having it do nothing.

Still to be done would be to depreciate the library functions and remove
tests relying on atomic invariants.