Experiment with alignment and padding.
This commit experiments with alignment and padding to STM structures
ensuring that nearly all structures are cache aligned to reduce false
conflicts with synchronization operations.  This appears to help TStruct
on one socket, but not later.  It also appears that there might be some
cache associativity problems with stacks in the default settings.