Attempt at having a separate type for RefAddr's that come from Alts.
I think in the end this breaks more assumptions then just having a
special cased mkVirtHeapOffsets for alts.  But I wasn't sure until I
tried.  Right now it emits the right code but cannot emit a readRef#
because of the use of a VoidRep id.  To fix this the ids need the right
rep for what they actually hold and mkVirtHeapOffsets will have to
change to know to do the right thing when the id's rep doesn't match the
layout.