Add dirty info tables for mutable constructors.
A dirty info table is made for each mutable constructor and the write barrier
for updating a ref will switch the closure to point to the dirty info table
if it is not already dirty and add the object to the mutable list.  During
garbage collection constructors with mutable fields are switched back to
clean.  The dirty and clean info tables refer to each other so the write
barrier and GC know which info table to switch to.

I don't think static mutable fields should be allowed, but I didn't do anything
at this level to prevent that.  A static info table is still generated.