Fix wrapper and teach lint about mutable case alts.
While the data constructor wrapper previously produced close to
the right code, it was clearly the wrong core.  It is now the
right core (for IO at least).

Core lint was failing due to finding the wrong type in case
alternatives.  Now the `dcAltRepType` field of `DataCon` exposes
the right type to check against in core lint.

Also added counter for mutable constructors on the `mutlist`.