Fix `dataConInstPat` and return Ref#'s correctly.
Clean up some code with data constructors and fix `dataConInstPat` which
needed the types for the alt, not the constructor when it has mutable
fields.  Also mirror unboxed tuple behavior in unarise of constructing
when encountering a null argument application.  I'm not sure if this is
quite right as we are making an unboxed tuple out of it, not the Ref
equivalent.  Some of unarise remains magical to me.

At this point I see strange behavior with worker-wrapper pass turned
on (I haven't investigated further)  and I get lint errors when
SpecConstr works on a mutable field constructor.  The rules it makes
are clearly wrong and it should do nothing for these constructors.