Add STM mutable constructors.
There is still a lot to be done here.  This extends the header for `STM`
(non-`IO` at this point) objects and emits code to initialize the `TStruct`
header.  Still needed is GC support (new object types so we can skip the
expanded header), separate `TRef` types (right now you can read and write
`STM` refs in `IO`), and probably a lot more.