Add mutable constrs with extended header for STM.
Allow for mutable constructors that have `STM` metadata in an
extended header.  This header makes the objects valid `TStruct`s
and only a couple of changes are needed to fix indexes from `Ref`s
and to correctly dirty with the right info pointer.  Note that
more complicated GC would be needed if we had pointers in the
header (e.g. watch lists).  We avoid pointers in the header with
bloom filter retry.

We still need to add separate `TRef` types to prevent access to
`STM` objects in `IO`.  There should also be a much more general
mechinism for setting the size and initalizing extended headers.
Right now this is just hard coded for what I need for STM, though
we store the size in the info table and use it at runtime (for GC).