Add unlifted unboxed mutable fields with RefU#.
Eventually we will want Ref# to be properly levity polymorphic, but
I don't think this version of GHC has all the bugs with that worked
out.  As a compromise we add RefU# which accepts only Int# as its
second type argument at the moment.  This is enough to demonstrate
the abilities we need.