Implement mutableArray.
This allows a mutable array fields at the end of a heap object
giving runtime allocated space of variable size.  Missing is
some design for intialization.  Here we minimally make GC happy
by initializing with self pointers.

With some design work we can add safety, but this should be close
to the performance we would see with that.